Changed score to an int type, Score was being set = SHRT_MIN in the code and then when an alignment forced a subtraction from this cell, we were getting wraparound which messed up the alignments
git-svn-id: file:///afs/nd.edu/user37/ccl/svn/cctools/trunk@930 a4d8336d-3463-0410-8bba-c098c45d37a8

+++ sand/src/matrix.h
@@ -9,7 +9,7 @@ See the file COPYING for details.
 
 struct cell
 {
-	short score;
+	int score;
 	short traceback;
 };
